Wraps content inside of header and footer
