+++ app/javascript/landing/Landing.js
@@ -1,107 +1,14 @@
 import React from "react";
-import AppBar from "@material-ui/core/AppBar";
-import IconButton from "@material-ui/core/IconButton";
-import MenuIcon from "@material-ui/icons/Menu";
-import CssBaseline from "@material-ui/core/CssBaseline";
-import Toolbar from "@material-ui/core/Toolbar";
-import Typography from "@material-ui/core/Typography";
-import { makeStyles } from "@material-ui/core/styles";
-import Link from "@material-ui/core/Link";
-import Button from "@material-ui/core/Button";
-
 import Hero from "./Hero";
 import HomeGrid from "./HomeGrid";
-import { UserContext } from "../context/UserContext";
 
-function Copyright() {
+const Landing = (props) => {
   return (
-    <Typography variant="body2" color="textSecondary" align="center">
-      {"Copyright © "}
-      <Link color="inherit" href="https://material-ui.com/">
-        Your Website
-      </Link>{" "}
-      {new Date().getFullYear()}
-      {"."}
-    </Typography>
+    <div>
+      <Hero handleSuccessfulAuth={props.handleSuccessfulAuth} />
+      <HomeGrid />
+    </div>
   );
-}
+};
 
-const useStyles = makeStyles((theme) => ({
-  icon: {
-    marginRight: theme.spacing(2),
-  },
-  root: {
-    flexGrow: 1,
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-  },
-  title: {
-    flexGrow: 1,
-  },
-  footer: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(6),
-  },
-}));
-
-export default function Landing(props) {
-  const classes = useStyles();
-  return (
-    <React.Fragment>
-      <CssBaseline />
-      <AppBar position="relative" className={classes.root}>
-        <Toolbar>
-          <IconButton
-            edge="start"
-            className={classes.menuButton}
-            color="inherit"
-            aria-label="menu"
-          >
-            <MenuIcon />
-          </IconButton>
-          <Typography
-            variant="h6"
-            color="inherit"
-            noWrap
-            className={classes.title}
-          >
-            Crystal Swap
-          </Typography>
-          <UserContext.Consumer>
-            {(value) => {
-              if (value === null) {
-                return <Button color="inherit">Login</Button>;
-              }
-              return (
-                <Button color="inherit" onClick={props.handleLogout}>
-                  Logout
-                </Button>
-              );
-            }}
-          </UserContext.Consumer>
-        </Toolbar>
-      </AppBar>
-      <main>
-        <Hero handleSuccessfulAuth={props.handleSuccessfulAuth} />
-        <HomeGrid />
-      </main>
-      {/* Footer */}
-      <footer className={classes.footer}>
-        <Typography variant="h6" align="center" gutterBottom>
-          Footer
-        </Typography>
-        <Typography
-          variant="subtitle1"
-          align="center"
-          color="textSecondary"
-          component="p"
-        >
-          Something here to give the footer a purpose!
-        </Typography>
-        <Copyright />
-      </footer>
-      {/* End footer */}
-    </React.Fragment>
-  );
-}
+export default Landing;
